Improve lua parser handling of unknown variables.
Previously the parser would assume any unknown variable was a new decl,
but if it wasn't the parser gave an rather unhelpful error. This
improves the error message by default, and also gives the caller a way
to fill undefined variables. The undefined variable is only a problem if
that caller is going to read that value, and that's not always the case.